// src/app/modules/movies/models/movies.model.ts
import {Genre} from './genres.model';
import {MovieOverview, MoviesCollection, ProductionCompany, SpokenLanguages} from './movie-overview.model';

export interface MovieDetails {
  adult: boolean;
  backdrop_path: string;
  belongs_to_collection: MoviesCollection;
  budget: number;
  genres: Genre[];
  homepage: string;
  id: number;
  imdb_id: string;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  production_companies: ProductionCompany[];
  release_date: string;
  revenue: number;
  runtime: number;
  spoken_languages: SpokenLanguages[];
  status: string;
  tagline: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
}

export interface MovieList {
  page: number;
  results: MovieOverview[];
  total_pages: number;
  total_results: number;
}